+++ flutter_easy_dialogs/lib/flutter_easy_dialogs.dart
@@ -1,7 +1,6 @@
 library flutter_easy_dialogs;
 
 import 'package:flutter/material.dart';
-import 'package:flutter_easy_dialogs/src/core/easy_dialog_manager.dart';
 import 'package:flutter_easy_dialogs/src/core/i_easy_dialog_manager_provider.dart';
 import 'package:flutter_easy_dialogs/src/overlay/easy_overlay_app_entry.dart';
 
@@ -36,27 +35,8 @@ class FlutterEasyDialogs extends StatelessWidget {
   static final _key = GlobalKey<EasyDialogsOverlayState>();
 
   /// Gets [IEasyDialogManagerProvider].
-  @visibleForTesting
   static IEasyDialogManagerProvider get provider =>
       _key.currentState!.dialogManagerProvider;
-
-  /// * Show:
-  ///
-  /// ```dart
-  /// FlutterEasyDialogs.use<MyDialogManager>().show(
-  ///       params: const EasyDialogManagerShowParams(
-  ///         content: Text('My custom manager'),
-  ///       ),
-  ///     );
-  ///);
-  /// ```
-  ///
-  ///* Hide:
-  ///
-  /// ```dart
-  /// FlutterEasyDialogs.use<MyDialogManager>().hide();
-  ///```
-  static M use<M extends EasyDialogManager>() => provider.get<M>();
 
   /// For using in [MaterialApp.builder].
   static const builder = _builder;