+++ flutter_easy_dialogs/lib/src/core/i_easy_dialog_manager_provider.dart
@@ -4,7 +4,7 @@ import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 ///
 /// Implementations of this interface are responsible for retrieving
 /// a particular instance of [EasyDialogManager] by type with the help of
-/// [get] method.
+/// [use] method.
 abstract class IEasyDialogManagerProvider {
   ///
   /// Retrieves an instance of [M], where [M] is a subtype of [EasyDialogManager].
@@ -15,5 +15,23 @@ abstract class IEasyDialogManagerProvider {
   /// If an instance of [M] has already been
   /// registered with the [IEasyDialogManagerRegistry],
   /// this method should return that instance.
-  M get<M extends EasyDialogManager>();
+  ///
+  ///
+  /// * Show:
+  ///
+  /// ```dart
+  /// FlutterEasyDialogs.provider.use<MyDialogManager>().show(
+  ///       params: const EasyDialogManagerShowParams(
+  ///         content: Text('My custom manager'),
+  ///       ),
+  ///     );
+  ///);
+  /// ```
+  ///
+  ///* Hide:
+  ///
+  /// ```dart
+  /// FlutterEasyDialogs.provider.use<MyDialogManager>().hide();
+  ///```
+  M use<M extends EasyDialogManager>();
 }