+++ full_screen_dialog_manager/lib/src/util/full_screen_dialog_controller_x.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 import 'package:full_screen_dialog_manager/src/manager/full_screen_dialog_manager.dart';
 
-extension FullScreenDialogControllerX on IEasyDialogManagerController {
+extension FullScreenDialogControllerX on IEasyDialogManagerProvider {
   Future<void> showFullScreen(FullScreenShowParams params) =>
       use<FullScreenDialogManager>().show(params: params);