+++ full_screen_dialog_manager/test/src/dismissible/gesture_test.dart
@@ -15,7 +15,7 @@ void main() {
       },
     ));
 
-    easyOverlayState.dialogManagerController.showFullScreen(
+    easyOverlayState.dialogManagerProvider.showFullScreen(
       const FullScreenShowParams(
         content: SizedBox(
           key: dialogKey,