+++ positioned_dialog_manager/lib/src/util/positioned_dialog_manager_controller_x.dart
@@ -1,9 +1,8 @@
-
 import 'package:flutter_easy_dialogs/flutter_easy_dialogs.dart';
 import 'package:positioned_dialog_manager/src/easy_dialog_position.dart';
 import 'package:positioned_dialog_manager/src/manager/positioned_dialog_manager.dart';
 
-extension PositionedDialogManagerControllerX on IEasyDialogManagerController {
+extension PositionedDialogManagerControllerX on IEasyDialogManagerProvider {
   Future<void> showPositioned(PositionedShowParams params) =>
       use<PositionedDialogManager>().show(params: params);