+++ positioned_dialog_manager/test/src/dismissible/gesture_test.dart
@@ -30,7 +30,7 @@ void main() {
     const position = EasyDialogPosition.top;
 
     unawaited(
-      easyOverlayState.dialogManagerController
+      easyOverlayState.dialogManagerProvider
           .showPositioned(const PositionedShowParams(
         dismissible: PositionedDismissible.gesture(),
         content: Text(