+++ positioned_dialog_manager/test/src/dismissible/none_test.dart
@@ -22,7 +22,7 @@ void main() {
     const position = EasyDialogPosition.top;
 
     unawaited(
-      easyOverlayState.dialogManagerController
+      easyOverlayState.dialogManagerProvider
           .showPositioned(const PositionedShowParams(
         dismissible: PositionedDismissible.none(),
         content: Text(