+++ positioned_dialog_manager/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+* Increased `FlutterEasyDialogs` dependency version.
+
 ## 1.0.0
 
 Initial release